requirements.txt
Babel==2.3.4
defusedxml==0.4.1
Django==1.10
django-activeurl==0.1.9
django-appconf==1.0.2
django-auditlog==0.4.0
django-autoslug==1.9.3
django-cities-light==3.3.0
django-ckeditor==5.0.3
django-classy-tags==0.7.2
django-debug-toolbar==1.4
django-filter==0.13.0
django-jsonfield==1.0.1
django-phonenumber-field==1.1.0
djangorestframework==3.3.3
iyzipay==1.0.21
lxml==3.6.0
Markdown==2.6.6
oauthlib==1.1.2
phonenumberslite==7.5.1
Pillow==3.3.0
progressbar2==3.10.0
PyJWT==1.4.0
python-utils==2.0.0
python3-openid==3.0.10
pytz==2016.6.1
reportlab==3.3.0
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.19
Unidecode==0.4.19
wheel==0.24.0

git+http://github.com/dogukankotan/python-social-auth.git

